+++ src/ScottPlot5/ScottPlot5/Plottables/CandlestickPlot.cs
@@ -52,7 +52,14 @@ public class CandlestickPlot(IOHLCSource data) : IPlottable
             limits = new AxisLimits(0, Data.GetOHLCs().Count, limits.Bottom, limits.Top);
         }
 
-        return limits;
+        List<OHLC> ohlcs = Data.GetOHLCs();
+        if (ohlcs.Count == 0)
+            return limits;
+
+        double left = ohlcs.First().DateTime.ToOADate() - ohlcs.First().TimeSpan.TotalDays / 2;
+        double right = ohlcs.Last().DateTime.ToOADate() + ohlcs.Last().TimeSpan.TotalDays / 2;
+
+        return new(left, right, limits.Bottom, limits.Top);
     }
 
     public void Render(RenderPack rp)
@@ -65,7 +72,7 @@ public class CandlestickPlot(IOHLCSource data) : IPlottable
             OHLC ohlc = ohlcs[i];
             bool isRising = ohlc.Close >= ohlc.Open;
             LineStyle lineStyle = isRising ? RisingLineStyle : FallingLineStyle;
-            FillStyle fillStlye = isRising ? RisingFillStyle : FallingFillStyle;
+            FillStyle fillStyle = isRising ? RisingFillStyle : FallingFillStyle;
 
             float top = Axes.GetPixelY(ohlc.High);
             float bottom = Axes.GetPixelY(ohlc.Low);
@@ -73,12 +80,11 @@ public class CandlestickPlot(IOHLCSource data) : IPlottable
             float center, left, right;
             if (Sequential == false)
             {
-                center = Axes.GetPixelX(ohlc.DateTime.ToNumber());
-                TimeSpan halfWidth = new((long)(ohlc.TimeSpan.Ticks * SymbolWidth / 2));
-                DateTime leftTime = ohlc.DateTime - halfWidth;
-                DateTime rightTime = ohlc.DateTime + halfWidth;
-                left = Axes.GetPixelX(leftTime.ToNumber());
-                right = Axes.GetPixelX(rightTime.ToNumber());
+                double centerNumber = ohlc.DateTime.ToNumber();
+                center = Axes.GetPixelX(centerNumber);
+                double halfWidthNumber = ohlc.TimeSpan.TotalDays / 2 * SymbolWidth;
+                left = Axes.GetPixelX(centerNumber - halfWidthNumber);
+                right = Axes.GetPixelX(centerNumber + halfWidthNumber);
             }
             else
             {
@@ -104,7 +110,7 @@ public class CandlestickPlot(IOHLCSource data) : IPlottable
 
             // rectangle
             SKRect rect = new(left, Math.Max(open, close), right, Math.Min(open, close));
-            fillStlye.ApplyToPaint(paint, rect.ToPixelRect());
+            fillStyle.ApplyToPaint(paint, rect.ToPixelRect());
             rp.Canvas.DrawRect(rect, paint);
         }
     }