+++ src/ScottPlot5/ScottPlot5/Primitives/OHLC.cs
@@ -9,6 +9,16 @@ public struct OHLC
     public DateTime DateTime { get; set; }
     public TimeSpan TimeSpan { get; set; }
 
+    public OHLC(double open, double high, double low, double close)
+    {
+        Open = open;
+        High = high;
+        Low = low;
+        Close = close;
+        DateTime = DateTime.MinValue;
+        TimeSpan = TimeSpan.FromDays(1);
+    }
+
     public OHLC(double open, double high, double low, double close, DateTime start, TimeSpan span)
     {
         Open = open;
@@ -17,6 +27,26 @@ public struct OHLC
         Close = close;
         DateTime = start;
         TimeSpan = span;
+    }
+
+    public readonly OHLC Clone()
+    {
+        return new OHLC(Open, High, Low, Close, DateTime, TimeSpan);
+    }
+
+    public readonly OHLC WithDate(DateTime dateTime)
+    {
+        return new OHLC(Open, High, Low, Close, dateTime, TimeSpan);
+    }
+
+    public readonly OHLC ShiftedBy(TimeSpan timeSpan)
+    {
+        return new OHLC(Open, High, Low, Close, DateTime + timeSpan, TimeSpan);
+    }
+
+    public readonly OHLC ShiftedBy(double delta)
+    {
+        return new OHLC(Open + delta, High + delta, Low + delta, Close + delta, DateTime, TimeSpan);
     }
 }